requirements.txt
asgiref==3.5.2
Brotli==1.0.9
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.4
pytz==2022.4
sqlparse==0.4.3
whitenoise==6.2.0
